implementazione sendMessages
Il bot risponde al messaggio con il contenuto del messaggio che invia l'utente

+++ TelegramAPI/TelegramAPI/src/TelegramAPI/Info.java
@@ -48,7 +48,7 @@ public class Info {
         this.c = c;
     }
 
-    public int getChatId()
+    public long getChatId()
     {
         return c.getId();
     }
